end experimental search. add image to start page.
