+++ ReactUI/src/components/index.js
@@ -1,7 +1,7 @@
 import DynamicForm from './DynamicForm';
 import MarvinEditor from './MarvinEditor';
 import PageSteps from './PageSteps';
-import SearchInput from './SearchInput';
+import SearchInput from '../search/hoc/SearchInput';
 import MainLayout from './MainLayout';
 import Loader from './Loader';
 import Error from './Error';
