+++ ReactUI/src/search/hoc/index.js
@@ -1,0 +1,3 @@
+import SearchInput from './SearchInput';
+
+export { SearchInput };
