+++ ReactUI/src/search/pages/HistoryPage.jsx
@@ -21,40 +21,43 @@ const H2 = styled.h2`
     padding-bottom: 30px;
 `;
 
-const HistoryPage = ({ histories, showResults, showValidate }) => (<div>
-  <H2>History result page</H2>
-  <div className="row">
-    { histories && histories.map(hist =>
-      (<div className="col-md-6">
-        <div className="thumbnail">
-          <Rigth>
-            <Tooltip placement="topLeft" title="Go to validate">
-              <Button
-                type="primary"
-                ghost
-                shape="circle"
-                icon="double-left"
-                size="large"
-                onClick={() => showValidate(hist.validateTaskId)}
-              />
-            </Tooltip>
-            <Tooltip placement="topLeft" title="Show result">
-              <Button
-                type="primary"
-                ghost
-                shape="circle"
-                icon="double-right"
-                size="large"
-                onClick={() => showResults(hist.resultTaskId)}
-              />
-            </Tooltip>
-          </Rigth>
-          <img src={hist.base64} />
-          <div><b>Selected model:</b> { hist.models.filter(m => m.model === hist.selectModel.model)[0].name }</div>
-        </div>
-      </div>))}
-  </div>
-</div>);
+const HistoryPage = ({ histories, showResults, showValidate }) => (
+  <div
+    style={{ padding: '50px 0', background: 'white' }}
+  >
+    <H2>History result page</H2>
+    <div className="row">
+      { histories && histories.map(hist =>
+        (<div className="col-md-6">
+          <div className="thumbnail">
+            <Rigth>
+              <Tooltip placement="topLeft" title="Go to validate">
+                <Button
+                  type="primary"
+                  ghost
+                  shape="circle"
+                  icon="double-left"
+                  size="large"
+                  onClick={() => showValidate(hist.validateTaskId)}
+                />
+              </Tooltip>
+              <Tooltip placement="topLeft" title="Show result">
+                <Button
+                  type="primary"
+                  ghost
+                  shape="circle"
+                  icon="double-right"
+                  size="large"
+                  onClick={() => showResults(hist.resultTaskId)}
+                />
+              </Tooltip>
+            </Rigth>
+            <img src={hist.base64} />
+            <div><b>Selected model:</b> { hist.models.filter(m => m.model === hist.selectModel.model)[0].name }</div>
+          </div>
+        </div>))}
+    </div>
+  </div>);
 
 const mapStateToProps = state => ({
   histories: state.histories,
