+++ ReactUI/src/search/pages/IndexPage.jsx
@@ -1,22 +1,13 @@
 import React from 'react';
-import { connect } from 'react-redux';
-import { SearchInput } from '../../components';
-import { URLS } from '../../config';
-import {
-  SAGA_DRAW_STRUCTURE,
-  SAGA_CREATE_TASK,
-} from '../core/constants';
 
-const mapStateToProps = () => ({
-  buttonURL: [
-    { name: 'Index page', url: URLS.INDEX },
-    { name: 'Info', url: URLS.INFO },
-    { name: 'Hisrory', url: URLS.HISTORY }],
-});
+const IndexPage = () => (<div>
+  <img
+    style={{
+      width: '100%',
+    }}
+    src="/static/images/1.png"
+    alt="start page image"
+  />
+</div>);
 
-const mapDispatchToProps = dispatch => ({
-  drawStructure: () => dispatch({ type: SAGA_DRAW_STRUCTURE }),
-  onSearchFormSubmit: data => dispatch({ type: SAGA_CREATE_TASK, data }),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(SearchInput);
+export default IndexPage;
