+++ ReactUI/src/search/pages/ResultPage.jsx
@@ -1,36 +1,14 @@
 import React, { Component } from 'react';
-import styled from 'styled-components';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Pagination, BackTop } from 'antd';
 import { ResultItem, ModalIncrease } from '../../components';
-import SearchInputView from './IndexPage';
+import { getRequest, getResults } from '../core/selectors';
 import {
   SAGA_CREATE_TASK,
   SAGA_INIT_RESULT_PAGE,
 } from '../core/constants';
 
-const Wrapper = styled.div`
-    padding-bottom: 20px;
-`;
-
-const CenterWrap = styled.div`
-    width: 100%;
-    padding-top: 20px;
-    padding-bottom: 40px;
-    text-align: center;
-`;
-
-const ResultWrapped = styled.div`
-    padding-top: 50px;
-`;
-
-const RightWrap = styled.div`
-    width: 100%;
-    padding-top: 20px;
-    padding-bottom: 40px;
-    text-align: right;
-`;
 
 class ResultPage extends Component {
   constructor(props) {
@@ -46,27 +24,23 @@ class ResultPage extends Component {
       results,
       showIncreaseModel,
       onSearchImg,
+      request: { loading, error },
     } = this.props;
 
-    return results && (
-      <div>
-        <Wrapper>
-          <SearchInputView />
-        </Wrapper>
-
-        <ResultWrapped>
-          {results && results.map((result, count) =>
-            (<ResultItem
-              count={count}
-              base64={result.base64}
-              onClickIcrease={() => showIncreaseModel(result.base64)}
-              result={result.models[0].results}
-              onSearchImage={() => onSearchImg(result.data)}
-            />),
-          )
-          }
-        </ResultWrapped>
-
+    return !loading && !error && results && (
+      <div
+        style={{ padding: '50px 0', background: 'white' }}
+      >
+        {results && results.map((result, count) =>
+          (<ResultItem
+            count={count}
+            base64={result.base64}
+            onClickIcrease={() => showIncreaseModel(result.base64)}
+            result={result.models[0].results}
+            onSearchImage={() => onSearchImg(result.data)}
+          />),
+        )
+        }
         <BackTop />
       </div>
     );
@@ -77,12 +51,9 @@ ResultPage.propTypes = {
 
 };
 
-ResultPage.defaultProps = {
-
-};
-
 const mapStateToProps = state => ({
-  results: state.results,
+  results: getResults(state),
+  request: getRequest(state),
 });
 
 const mapDispatchToProps = dispatch => ({
